backend/requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
